requirements.txt
git+https://github.com/Rapptz/discord.py@rewrite#egg=discord.py

asyncpg
click
emoji>=0.5
lru_dict
lxml
more-itertools>=4.0
parsedatetime
pillow
psutil
python-dateutil
uvloop